+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+The format conforms to [Keep a Changelog](http://keepachangelog.com/).
+
+## [Unreleased]
+### Changed
+- Set `checkLoops` to `false` for rule `no-constant-condition`
 
 ## 1.0.0 - 2016-06-25
 ### Added
 - Initial rule set
+
+[Unreleased]: https://github.com/MitMaro/eslint-config-mitmaro/compare/v1.0.0...HEAD